fix: 🐛 move homepage styled components out of pages directory
